requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
streamlit
plotly
python-dotenv
langchain
langchain-openai
openai